requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
langchain==0.3.0
sentence-transformers==3.1.1
faiss-cpu==1.8.0
python-dotenv==1.0.1
transformers==4.44.2
torch==2.4.1
pdfminer.six==20231228
pydantic==2.9.2